+++ src/video/track_publisher.cpp
@@ -48,7 +48,10 @@ void tryUnpublish(RoomConnection & connection, const std::shared_ptr<livekit::Lo
 
 livekit::TrackPublishOptions publishOptionsWithFrameMetadata(livekit::TrackPublishOptions options)
 {
-  options.packet_trailer_features.user_timestamp = true;
+  livekit::FrameMetadataFeatures metadata_features;
+  metadata_features.user_timestamp = true;
+  metadata_features.frame_id = false;
+  options.frame_metadata_features = metadata_features;
   return options;
 }